requirements.txt



langchain



streamlit


pypdf                 # For loading PDF files
langchain-community   # Contains the PyPDFLoader
langchain_text_splitters # For the chunking logic


chromadb


sentence-transformers  # The underlying model engine
langchain-huggingface  # The LangChain integration for it


langchain-groq         # The LangChain integration for the Groq API


python-dotenv          # For managing API keys in the .env file